Provide keepAliveTimeout setting for the app
This adds a new setting `settings:keep_alive_timeout` or
`TX__SETTINGS__KEEP_ALIVE_TIMEOUT_SEC` that if it is different than the
default value `''` adds a keepAliveTimeout setting to the node server.

This can be really helpful in situations that there is a load balancer
in front of the application and this setting needs to be a bit higher
than the load balancer's idle timeout.